requirements.txt
aiogram>=3.15.0
aiosqlite>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
